feat: add --nan support to 'e test'
