+++ src/utils/logging.js
@@ -21,7 +21,11 @@ const color = {
 };
 
 function fatal(e) {
-  console.error(`${color.err} ${e.stack ? e.stack : e.message}`);
+  if (typeof e === 'string') {
+    console.error(`${color.err} ${e}`);
+  } else {
+    console.error(`${color.err} ${e.stack ? e.stack : e.message}`);
+  }
   process.exit(1);
 }
